Replace old examples
